Adding stand alone CISM documentation
I am adding a version of the stand alone CISM documentation to be
available for the CISM 2.1 release.
Note: although this version of the documentation contains most of
the major changes, some cleanup remains but are not critical for the
comprehension or understanding of the model and its current features.